pythonservers/ADK/requirements.txt
# ADK Server Dependencies for Hugging Face Deployment

# Core Framework
Flask==3.0.3
Flask-CORS==4.0.0

# Google ADK & Vertex AI
google-adk>=0.1.0
google-cloud-aiplatform>=1.35.0
vertexai>=0.42.0
google-generativeai>=0.5.0

# Web & HTTP
requests==2.31.0
Werkzeug==3.0.1

# Data Processing
beautifulsoup4==4.12.2
lxml==4.9.3

# Utilities
python-dotenv==1.0.0
Jinja2==3.1.2

# Optional: For enhanced performance
gunicorn==21.2.0

# Optional: For Playwright if needed (scraping)
# playwright==1.48.0  # Uncomment if you need browser automation
